feat: remove image_to_build folder use finally

+++ app/server/instance_server.py
@@ -2,7 +2,7 @@ import os.path
 import shutil
 from app import proto
 from ..utils import docker_utils
-from ..mq import push_message
+# from ..mq import push_message
 from git import Repo
 import json
 from google.protobuf.json_format import MessageToDict
@@ -16,23 +16,35 @@ class InstanceService(proto.instance_pb2_grpc.InstanceServiceServicer):
         return proto.instance_pb2.CreateInstanceResponse(instance_id=container.id, instance_name=container.name)
 
     def CreateImage(self, request, context):
-        image_to_build_path = "image_to_build"
-        Repo.clone_from("https://github.com/TREXA23187/flask_demo", image_to_build_path)
+        try:
+            if os.path.exists("image_to_build"):
+                shutil.rmtree("image_to_build")
+                
+            image_to_build_path = "image_to_build"
+            Repo.clone_from("https://github.com/TREXA23187/flask_demo", image_to_build_path)
 
-        with open(os.path.join(image_to_build_path, "data/data.csv"), 'wb') as data_file:
-            data_file.write(request.data_file)
+            if request.data_file:
+                with open(os.path.join(image_to_build_path, "data/data.csv"), 'wb') as data_file:
+                    data_file.write(request.data_file)
 
-        with open(os.path.join(image_to_build_path, "model/enter.py"), 'wb') as model_file:
-            model_file.write(request.model_file)
+            if request.model_file:
+                with open(os.path.join(image_to_build_path, "model/enter.py"), 'wb') as model_file:
+                    model_file.write(request.model_file)
 
-        with open(os.path.join(image_to_build_path, "config.json"), 'w') as config_file:
-            json.dump(MessageToDict(request.image_config), config_file)
+            if request.image_config:
+                with open(os.path.join(image_to_build_path, "config.json"), 'w') as config_file:
+                    json.dump(MessageToDict(request.image_config), config_file)
 
-        image = docker_utils.build_image(image_to_build_path, f"{request.repository}:{request.tag}")
+            image = docker_utils.build_image(image_to_build_path, f"{request.repository}:{request.tag}")
 
-        push_message(request.repository.encode('utf-8'))
-        shutil.rmtree("image_to_build")
-        return proto.instance_pb2.CreateImageResponse(image_id=image.id, image_size=image.attrs['Size'])
+            # push_message(request.repository.encode('utf-8'))
+            return proto.instance_pb2.CreateImageResponse(image_id=image.id, image_size=image.attrs['Size'])
+        except Exception as e:
+            print("create image error: ", e)
+
+        finally:
+            if os.path.exists("image_to_build"):
+                shutil.rmtree("image_to_build")
 
     def DeleteImage(self, request, context):
         success = docker_utils.delete_image(request.repository, request.tag)
